backend/fastapi/requirements.txt
fastapi
fastapi-cli
requests
uvicorn
pillow
paddlepaddle 
paddleocr
colorama
python-multipart